% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdim-package.R
\docType{package}
\name{gdim-package}
\alias{gdim}
\alias{gdim-package}
\title{gdim: Graph Dimensionality}
\description{
Estimate the graph dimensionality using thecross-validated 
    eigenvalues. The cross-validated eigenvalues are estimated by splitting 
    a sample graph into two, the training and the test graph. The training graph
    is use to estimate sample eigen eigenvectors, and the test graph is use to
    evaluate the correlation between the sample eigenvectors and the eigenspace
    of the test graph (thus the population eigenspace). The test statistics for
    graph dimensionality follows a simple central limit theorem. The p-values 
    are provided for each hypothesized graph dimensionality sequentially.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RoheLab/gdim}
  \item Report bugs at \url{https://github.com/RoheLab/gdim/issues}
}

}
\author{
\strong{Maintainer}: Fan Chen \email{fan.chen@wisc.edu} (\href{https://orcid.org/0000-0003-4508-6023}{ORCID})

Authors:
\itemize{
  \item Karl Rohe \email{karlrohe@stat.wisc.edu}
}

}
\keyword{internal}
