% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigcv.R
\name{glaplacian}
\alias{glaplacian}
\title{Graph Laplacian}
\usage{
glaplacian(A, regularize = TRUE)
}
\arguments{
\item{A}{the adjacency matrix, will be coerced to a \code{Matrix} class.}

\item{regularize}{\code{logical(1)}, return a regularized symmetric Laplacian.
The default is \code{TRUE}. This is ignored if \code{laplacian=FALSE}.}
}
\description{
Given an \code{m} by \code{n} graph adjacency matrix \code{A}, calculate the symmetric
graph Laplacian.
}
