% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphdim-package.R
\docType{package}
\name{graphdim-package}
\alias{graphdim}
\alias{graphdim-package}
\title{graphdim: Estimate the Dimension of a Graph}
\description{
Uses Karl and Sebastien's trick to estimate the dimension
    of a graph using a nonparametric resampling test.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexpghayes/graphdim}
  \item Report bugs at \url{https://github.com/alexpghayes/graphdim/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (0000-0002-4985-5160)

Authors:
\itemize{
  \item Karl Rohe \email{KarlRohe@wisc.edu}
}

}
\keyword{internal}
