% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigcv.R
\name{gspectral}
\alias{gspectral}
\title{Graph Spectral Decomposition}
\usage{
gspectral(A, k, ...)
}
\arguments{
\item{A}{the adjacency matrix, will be coerced to a \code{Matrix} class.}

\item{k}{\code{integer(1)}, number of eigenvalues.}

\item{...}{additonal parameters to pass into \link[RSpectra:svds]{RSpectra::svds}.}
}
\value{
a list (see also \link[RSpectra:svds]{RSpectra::svds}):
\item{u}{an \code{m} by \code{k} matrix whose columns contain the left singular vectors.}
\item{v}{an \code{n} by \code{k} matrix whose columns contain the right singular vectors.}
}
\description{
Given ab \code{m} by \code{n} graph matrix \code{A}, find the largest \code{k} singular values
and the corresponding singular vectors.
}
