% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horn.R
\name{horn}
\alias{horn}
\title{Estimate the rank of a matrix with the Kaiser-Guttman Rule}
\usage{
horn(X, ...)
}
\arguments{
\item{X}{A matrix.}
}
\value{
The estimated rank of the matrix, as an integer.
}
\description{
That is, set \code{k} to the number of components of the
correlation matrix with eigenvalue greater than one.
}
\examples{

X <- iris[, 1:4]
kaiser(X)

}
