% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Calculate the degree normalization of an adjacency matrix}
\usage{
normalize(A, tau_row = NULL, tau_col = NULL)
}
\arguments{
\item{A}{An adjacency matrix. Either a \code{\link[=matrix]{matrix()}} or a \code{\link[Matrix:Matrix]{Matrix::Matrix()}}.}

\item{tau_row}{Out degree regularization term. Should be positive.
Defaults to \code{NULL}, in which case we use the mean out degree.}

\item{tau_col}{In degree regularization term. Should be positive.
Defaults to \code{NULL}, in which case we use the mean in degree.}
}
\description{
Calculate the degree normalization of an adjacency matrix
}
