% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigcv.R
\name{plot.eigcv}
\alias{plot.eigcv}
\title{Plot \code{eigcv}}
\usage{
\method{plot}{eigcv}(x, type = "z", threshold = 2, verbose = FALSE, ...)
}
\arguments{
\item{x}{an \code{eigcv} object.}

\item{type}{either "z" or "p" to specify the y-axis of the plot.
If "z", plot the test statistics (asymptotic z score) for each k.
If "p", plot the log10 transformed p-values.}

\item{threshold}{\code{numeric(1)}, cut-off of p-value (in log10), default to 2.}

\item{verbose}{\code{logical(1)}, whether to print out the summary table.}

\item{...}{ignored.}
}
\value{
Plot an \code{eigcv} object.
}
\description{
Plot \code{eigcv}
}
