% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigcv.R
\name{print.eigcv}
\alias{print.eigcv}
\title{Print \code{eigcv}}
\usage{
\method{print}{eigcv}(x, ...)
}
\arguments{
\item{x}{an \code{eigcv} object.}

\item{...}{additional input to generic \link{print}.}

\item{verbose}{\code{logical(1)}, whether to print out the summary table.}
}
\value{
Print an \code{eigcv} object interactively.
}
\description{
Print \code{eigcv}
}
