% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgof.R
\name{print.stgof}
\alias{print.stgof}
\title{Print \code{stgof}}
\usage{
\method{print}{stgof}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{an \code{stgof} object.}

\item{verbose}{\code{logical(1)}, whether to print out the summary table.}

\item{...}{additional input to generic \link{print}.}
}
\value{
Print an \code{stgof} object interactively.
}
\description{
Print \code{stgof}
}
