% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_permute.R
\name{row_permute}
\alias{row_permute}
\title{Estimate the rank of a graph adjacency matrix with row permutation}
\usage{
row_permute(X, k_max = min(ncol(X), 1000), ...)
}
\arguments{
\item{X}{A matrix. Sparse matrices are automatically
coerces to dense matrices, so be careful.}
}
\value{
The estimated rank of the matrix, as an integer.
}
\description{
Estimate the rank of a graph adjacency matrix with row permutation
}
\examples{

set.seed(27)

n <- 1000
k <- 5

B <- matrix(runif(k * k), nrow = k, ncol = k) # mixing probabilities

theta <- round(rlnorm(n, 2)) # degree parameter
pi <- c(1, 2, 4, 1, 1) / 5 # community membership

A <- dcsbm(theta, pi, B, avg_deg = 50)
kaiser(A)

X <- iris[, 1:4]
kaiser(X)

}
