% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgof.R
\name{RQ}
\alias{RQ}
\title{Refitted Quadrilateral}
\usage{
RQ(A, label)
}
\arguments{
\item{A}{a \code{matrix} or \code{Matrix}, the adjacency matrix of an \emph{undirected} network.}

\item{label}{a vector of \code{1:K}, the community labels of all nodes.}
}
\value{
The RQ statistics.
}
\description{
Calculate the Refitted Quadrilateral (RQ) statistics.
}
\references{
Jin, Jiashun. Fast community detection by SCORE. Ann. Statist. 43 (2015), no. 1, 57--89. doi:10.1214/14-AOS1265. \url{https://projecteuclid.org/euclid.aos/1416322036}
}
