% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgof.R
\name{SCORE}
\alias{SCORE}
\title{Spectral Clustering On Ratios-of-Eigenvectors}
\usage{
SCORE(
  A,
  K,
  truncate = log(nrow(A)),
  iter.max = 1000,
  nstart = 1000,
  algorithm = "Hartigan-Wong",
  trace = FALSE
)
}
\arguments{
\item{A}{a \code{matrix} or \code{Matrix}, the adjacency matrix of an \emph{undirected} network.}

\item{K}{\code{integer(1)}, number of dimension.}

\item{truncate}{\code{numeric(1)}, truncate the eigen-ratio (in absolute value)
matrix, default to \code{log(nrow(A))}. Set this to \code{0} if don't want truncation.}

\item{iter.max}{the maximum number of iterations allowed.}

\item{nstart}{if \code{centers} is a number, how many random sets
    should be chosen?}

\item{algorithm}{character: may be abbreviated.  Note that
    \code{"Lloyd"} and \code{"Forgy"} are alternative names for one
    algorithm.}

\item{trace}{logical or integer number, currently only used in the
    default method (\code{"Hartigan-Wong"}): if positive (or true),
    tracing information on the progress of the algorithm is
    produced.  Higher values may produce more tracing information.}
}
\value{
a vector of \code{1:K}, the community labels of all nodes.
}
\description{
Perform Spectral Clustering On Ratios-of-Eigenvectors (SCORE).
}
\references{
Jin, Jiashun. Fast community detection by SCORE. Ann. Statist. 43 (2015), no. 1, 57--89. doi:10.1214/14-AOS1265. \url{https://projecteuclid.org/euclid.aos/1416322036}
}
