% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign-flip.R
\name{sf_distribution}
\alias{sf_distribution}
\title{Sampling distribution of row permutation test lambda}
\usage{
sf_distribution(X, k, B = 1000)
}
\arguments{
\item{B}{}
}
\description{
Sampling distribution of row permutation test lambda
}
\examples{

library(Matrix)
library(tidyverse)

A <- rsparsematrix(1000, 1000, 0.1)

# symmetrize
A <- sign(t(A) + A)

sampling_dist <- sf_distribution(A, 10, B = 200)

sampling_dist \%>\%
  as_tibble() \%>\%
  mutate(B = row_number()) \%>\%
  gather(lambda, value, -B) \%>\%
  ggplot(aes(value)) +
  geom_histogram() +
  facet_wrap(~lambda) +
  theme_minimal()

}
