% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigcv.R
\name{split_graph}
\alias{split_graph}
\title{Graph Splitting}
\usage{
split_graph(A, test_portion = 0.1)
}
\arguments{
\item{A}{the adjacency matrix, will be coerced to a \code{Matrix} class.}

\item{test_portion}{\code{numeric(1)}, the proportion of leave-out edges (into the test set).
Must take value from (0,1).}
}
\value{
A list of two \code{Matrix} elements:
\item{train}{the subgraph with \code{1-test_portion} portion of edges.}
\item{test}{the subgraph with \code{test_portion} portion of edges.}
}
\description{
Randomly split edges of graph into two, proportional to \code{1-test_portion} and \code{test_portion}.
}
