% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgof.R
\name{StGoF}
\alias{StGoF}
\title{Stepwise Goodness-of-Fit}
\usage{
StGoF(A, K_max, alpha = 0.05)
}
\arguments{
\item{A}{a \code{matrix} or \code{Matrix}, the adjacency matrix of an \emph{undirected} network.}

\item{K_max}{\code{integer(1)}, number of communities.}

\item{alpha}{\code{numeric(1)}, significance level, default to 0.05.
Set \code{alpha=1} if you want to compute all the statistics from 1 to K_max.}
}
\value{
a \code{data.table} with three columns:
\item{k}{tested graph dimension.}
\item{rq.stat}{the refitted quadrilateral test statistics.}
\item{p.value}{test p-value (two-sided, no adjustment).}
}
\description{
Perform the stepwise goodness-of-fit (StGoF) tests of graph dimension.
The tests are run from k=2, 3, ..., K_max separately.
Given a significance level (\code{alpha}), the infered graph dimension is
determined by the first test (i.e., smallest k) that is accepted.
}
\references{
Jin, Jiashun. Fast community detection by SCORE. Ann. Statist. 43 (2015), no. 1, 57--89. doi:10.1214/14-AOS1265. \url{https://projecteuclid.org/euclid.aos/1416322036}
}
