% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgof.R
\name{z2Z}
\alias{z2Z}
\title{From cluster label to cluster membership matrix}
\usage{
z2Z(z)
}
\arguments{
\item{z}{a \code{integer} vector, cluster label 1,2,...}
}
\value{
a binary \code{sparseMatrix} of cluster membership.
}
\description{
From cluster label to cluster membership matrix
}
