% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigcv.R
\name{eigcv}
\alias{eigcv}
\title{Edge Bootstrapping and Splitting}
\usage{
eigcv(
  A,
  k_max,
  ...,
  num_bootstraps = 10,
  test_portion = 0.1,
  alpha = 0.05,
  ptol = .Machine$double.eps,
  method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"),
  laplacian = FALSE,
  regularize = TRUE
)
}
\arguments{
\item{A}{The adjacency matrix of graph. Must be non-negative integer valued.}

\item{k_max}{\code{integer(1)}, number of eigenvectors to compute.}

\item{num_bootstraps}{\code{integer(1)}, number of graph bootstraps, default to 10.
Graph bootstrapping is to account for the randomness in graph splitting,
rather than obtaining any statistic (as a traditional num_bootstraps does).
Hence, a small number (e.g., 3~10) of bootstraps usually suffices.
If \code{num_bootstraps>1}, the test statistics will be averaged across bootstraps
and the p-values will be calculated based on the averaged statistics.}

\item{test_portion}{TODO}

\item{alpha}{Significance level of each test, defaults to \code{0.05}.
This is used to cut off the dimension estimation.}

\item{ptol}{\code{numeric(1)}, the tolerance of minimal p-value.}

\item{method}{correction method, a \code{\link{character}} string.
    Can be abbreviated.}

\item{laplacian}{\code{logical(1)}, use the normalized and regularized adjacency
matrix (i.e. L)
This option is experimental and should be used with caution.}

\item{regularize}{TODO}
}
\value{
A \code{eigcv} object, which contains:
\item{estimated_dimension}{inferred graph dimension.}
\item{summary}{summary table of the tests.}
\item{num_bootstraps}{number of bootstraps performed.}
\item{test_portion}{graph splitting probability used.}
\item{alpha}{significance level of each test.}
}
\description{
Estimate the graph dimension via eigenvalue cross-validation (EigCV).
A graph has dimension \code{k} if the first \code{k} eigenvectors of its adjacency
matrix are correlated with its population eigenspace, and the others are not.
Edge bootstrapping sub-samples the edges of the graph (without replacement).
Edge splitting separates the edges into a training part and a testing part.
}
\examples{

library(fastRG)

set.seed(27)

B <- matrix(0.1, 5, 5)
diag(B) <- 0.3

model <- sbm(
  n = 1000,
  k = 5,
  B = B,
  expected_degree = 40,
  poisson_edges = FALSE,
  allow_self_loops = FALSE
)

A <- sample_sparse(model)

eigcv_result <- eigcv(A, k_max = 10)
eigcv_result

}
