% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigcv.R
\name{eigcv}
\alias{eigcv}
\title{Edge Bootstrapping and Splitting}
\usage{
eigcv(
  A,
  k_max,
  bootstrap = 10,
  split = 0.1,
  alpha = 0.05,
  meta = "z",
  ptol = 1e-32,
  correct = "none",
  laplacian = TRUE,
  regularize = TRUE,
  align = FALSE,
  trace = FALSE
)
}
\arguments{
\item{A}{the adjacency matrix, will be coerced to a \code{Matrix} class.}

\item{k_max}{\code{integer(1)}, number of eigenvectors to compute.}

\item{bootstrap}{\code{integer(1)}, number of graph bootstraps, default to 10.
Graph bootstrapping is to account for the randomness in graph splitting,
rather than obtaining any statistic (as a traditional bootstrap does).
Hence, a small number (e.g., 3~10) of bootstraps usually suffices.
If \code{bootstrap>1}, the test statistics will be averaged across bootstraps
and the p-values will be calculated based on the averaged statistics.}

\item{split}{\code{numeric(1)}, the proportion of leave-out edges (into the test set).
Must take value from (0,1).}

\item{alpha}{\code{numeric(1)}, significance level of each test, default to 0.05.}

\item{meta}{either "z" or "p", how to aggregate bootstrap results.
If "z", will take the average of test statistics.
If "p", will perform meta analysis using \link[metap:sump]{metap::sump}.}

\item{ptol}{\code{numeric(1)}, the tolerance of minimal p-value.
This is only useful when \code{meta=="p"} (e.g., "bonferroni", "fdr").}

\item{correct}{\code{character(1)}, correction method for multiple testing.
This is recommended if \code{k_max} is large. See \link{p.adjust} for the acceptable
values of \code{method}. If \code{correct} is not "none", the returned summary table
will have an extra column called \code{padj}.}

\item{laplacian}{\code{logical(1)}, use the (symmetric) graph Laplacian.}

\item{regularize}{\code{logical(1)}, return a regularized symmetric Laplacian.
The default is \code{TRUE}. This is ignored if \code{laplacian=FALSE}.}

\item{align}{\code{logical(1)}, align the trained singular vectors with those of
the full graph. The default is \code{FALSE}.
This option is experimental and does not guarantee any improvement.}

\item{trace}{\code{logical(1)}, for diagnostic use. If \code{TRUE}, return additionally
\code{stats}, the test statistics of every bootstrap, and \code{pvals}, the eigen
vectors of the (last, if \code{bootstrap>1}) training graph.}
}
\value{
A \code{eigcv} object, which contains:
\item{inference}{inferred graph dimension.}
\item{summary}{summary table of the tests.}
\item{bootstrap}{number of bootstraps performed.}
\item{split}{graph splitting probability used.}
\item{alpha}{significance level of each test.}
\item{trace}{only if \code{trace=TRUE}, the spectral decomposition of the traning graph.}
}
\description{
Estimate the graph dimension via eigenvalue cross-validation (EigCV).
A graph has dimension \code{k} if the first \code{k} eigenvectors of its adjacency
matrix are correlated with its population eigenspace, and the others are not.
Edge bootstrapping sub-samples the edges of the graph (without replacement).
Edge splitting separates the edges into a training part and a testing part.
}
