% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdim-package.R
\docType{package}
\name{gdim-package}
\alias{gdim}
\alias{gdim-package}
\title{gdim: Estimate Graph Dimension using Cross-Validated Eigenvalues}
\description{
Cross-validated eigenvalues are estimated by splitting a graph into two parts, the training and the test graph. The training graph is used to estimate eigenvectors, and the test graph is used to evaluate the correlation between the training eigenvectors and the eigenvectors of the test graph. The correlations follow a simple central limit theorem that can be used to estimate graph dimension via hypothesis testing, see Chen et al. (2021) \href{https://arxiv.org/abs/2108.03336}{arXiv:2108.03336} for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RoheLab/gdim}
  \item \url{https://rohelab.github.io/gdim/}
  \item Report bugs at \url{https://github.com/RoheLab/gdim/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Fan Chen \email{fchen365@gmail.com} (\href{https://orcid.org/0000-0003-4508-6023}{ORCID})
  \item Karl Rohe \email{karlrohe@wisc.edu}
}

}
\keyword{internal}
