% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigcv.R
\name{plot.eigcv}
\alias{plot.eigcv}
\title{Plot \code{eigcv}}
\usage{
\method{plot}{eigcv}(x, type = c("z", "A", "L"), threshold = 2, ...)
}
\arguments{
\item{x}{an \code{eigcv} object.}

\item{type}{either "z", "A", or "L" to specify the y-axis of the plot.
If "z", plot the test statistics (asymptotic z score) for each k.
If "A", plot x'Ax for each eigenvector x.
If "L", plot x'Lx for each eigenvector x.}

\item{threshold}{\code{numeric(1)}, cut-off of p-value (in log10), default to 2.}

\item{...}{ignored.}
}
\value{
Plot an \code{eigcv} object.
}
\description{
Plot \code{eigcv}
}
